% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mc}
\alias{mc}
\title{Monte Carlo estimated covariance, CUDA accelerated}
\usage{
mc(x, s)
}
\arguments{
\item{x}{a vector of distances}

\item{s}{a vector of random samples from the estimated spectral density}
}
\value{
a vector
}
\description{
Monte Carlo estimated covariance, CUDA accelerated
}
